requirements.txt
openai-whisper
pydub
ffmpeg-python

scikit-learn
joblib
numpy

hmmlearn
joblib

textblob

scikit-learn
joblib

transformers
torch

TTS

noisereduce
librosa
soundfile

librosa
numpy
speechbrain
torch

